+++ packaging/windows/installer.go
@@ -13,7 +13,7 @@ import (
 	"strings"
 )
 
-const version = "1.7.4"
+const version = "1.7.5"
 
 //go:embed payload.zip
 var payload []byte